+++ RELEASES.md
@@ -5,6 +5,14 @@ modification, are permitted in any medium without royalty provided the
 copyright notice and this notice are preserved. This file is offered
 as-is, without any warranty. -->
 
+# Version 1.23.1 (unreleased)
+
+  * Bug fix: comparison of signed and unsigned numbers where the signed number
+    was wider than the unsigned number was truncating bits from the signed
+    number ([issue 57]).
+
+[issue 57]: https://gitlab.com/tspiteri/fixed/-/issues/57
+
 # Version 1.23.0 (2023-02-06)
 
   * String parsing of binary, octal and hexadecimal strings now supports base-2